Changes to refactor js code with utilities subdirectory, fix event listeners in static trees so that they show when user logs in directly to page, stop code from running multiple times if already logged in when accessing profile pages and person lists
