+++ docs/assets/js/table.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";var e={};e.d=(t,o)=>{for(var n in o)e.o(o,n)&&!e.o(t,n)&&Object.defineProperty(t,n,{enumerable:!0,get:o[n]})},e.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),window.addEventListener("DOMContentLoaded",(function(){!function(){for(var e=document.getElementsByTagName("th"),t=0;t<e.length;t++)e[t].classList.add("sorted-none");const o=document.getElementById("searchInput"),n=document.getElementById("myTable").querySelectorAll("tr:not(:first-child)");function r(e,t){let o;return function(...n){const r=this;clearTimeout(o),o=setTimeout((()=>e.apply(r,n)),t)}}function s(e){return e.toLowerCase().replace(/[^a-z0-9]+/g," ").trim().split(" ").filter((e=>e.length>2)).join(" ")}function a(e){return e.toLowerCase().replace(/[^a-z0-9]+/g," ").trim().split(" ")}o.value="",n.forEach((e=>{e.dataset.keywords=s(e.textContent)})),o.addEventListener("keyup",r((function(){const e=a(o.value);for(const t of n){const o=t.dataset.keywords,n=e.map((e=>o.includes(e))).every((e=>!0===e));t.style.display=n?"":"none"}}),100))}()})),window.addEventListener("pageshow",(function(e){searchInput.value="",document.getElementById("myTable").querySelectorAll("tr:not(:first-child)").forEach((e=>{e.style.display=""}))})),window.sortTable=function(e){const t=document.getElementById("myTable"),o=t.querySelectorAll("th"),n=Array.from(t.querySelectorAll("tr:not(:first-child)")),r=o[e],s="asc"===(r.dataset.sortDir||"none")?"desc":"asc";o.forEach((e=>{e.classList.remove("sorted-asc","sorted-desc","sorted-none"),e.dataset.sortDir="none",e.classList.add("sorted-none")})),r.classList.remove("sorted-none"),r.classList.add(`sorted-${s}`),r.dataset.sortDir=s,n.sort(((t,o)=>{const n=t.cells[e].textContent.trim(),r=o.cells[e].textContent.trim(),a=isNaN(n)?n.localeCompare(r):parseFloat(n)-parseFloat(r);return"asc"===s?a:-a}));for(const e of n)t.tBodies[0].appendChild(e)}})();
+(()=>{"use strict";window.addEventListener("DOMContentLoaded",(function(){!function(){for(var e=document.getElementsByTagName("th"),t=0;t<e.length;t++)e[t].classList.add("sorted-none");const o=document.getElementById("searchInput"),s=document.getElementById("myTable").querySelectorAll("tr:not(:first-child)");o.value="",s.forEach((e=>{e.dataset.keywords=e.textContent.toLowerCase().replace(/[^a-z0-9]+/g," ").trim().split(" ").filter((e=>e.length>2)).join(" ")})),o.addEventListener("keyup",function(e,t){let o;return function(...s){const n=this;clearTimeout(o),o=setTimeout((()=>e.apply(n,s)),t)}}((function(){const e=o.value.toLowerCase().replace(/[^a-z0-9]+/g," ").trim().split(" ");for(const t of s){const o=t.dataset.keywords,s=e.map((e=>o.includes(e))).every((e=>!0===e));t.style.display=s?"":"none"}}),100))}()})),window.addEventListener("pageshow",(function(e){searchInput.value="",document.getElementById("myTable").querySelectorAll("tr:not(:first-child)").forEach((e=>{e.style.display=""}))})),window.sortTable=function(e){const t=document.getElementById("myTable"),o=t.querySelectorAll("th"),s=Array.from(t.querySelectorAll("tr:not(:first-child)")),n=o[e],r="asc"===(n.dataset.sortDir||"none")?"desc":"asc";o.forEach((e=>{e.classList.remove("sorted-asc","sorted-desc","sorted-none"),e.dataset.sortDir="none",e.classList.add("sorted-none")})),n.classList.remove("sorted-none"),n.classList.add(`sorted-${r}`),n.dataset.sortDir=r,s.sort(((t,o)=>{const s=t.cells[e].textContent.trim(),n=o.cells[e].textContent.trim(),a=isNaN(s)?s.localeCompare(n):parseFloat(s)-parseFloat(n);return"asc"===r?a:-a}));for(const e of s)t.tBodies[0].appendChild(e)}})();
