+++ docs/assets/js/tree.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";function e(t,n,i){const a=n.querySelectorAll(":scope > li"),o=a[a.length-1],r=o.getBoundingClientRect().height/2,s=o.getBoundingClientRect(),c=t.getBoundingClientRect(),l=s.bottom-c.bottom-r;n.style.setProperty("--ul-before-margin",`${l}px`);const d=n.parentElement;if(i&null!==d&"UL"===d.tagName){const t=d.previousElementSibling;null!==t&&e(t,d,!0)}}function t(e){for(var n=e.getElementsByClassName("nested"),i=0;i<n.length;i++)n[i].classList.remove("active"),t(n[i]);var a=e.getElementsByClassName("collapsible");for(i=0;i<a.length;i++)a[i].classList.remove("active")}function n(n){n.classList.toggle("active");const i=n.parentElement.parentElement.parentElement,a=i.nextElementSibling;a.classList.toggle("active"),e(i,a,!0),n.classList.contains("active")?n.textContent="":(n.textContent="",t(a))}var i=!1;async function a(){var e={email:"",user_id:""};try{await new Promise(((e,t)=>{let n=0;const i=setInterval((()=>{"undefined"!=typeof MemberSpace?(clearInterval(i),e()):(n+=100,n>=1e4&&(clearInterval(i),t("MemberSpace did not become available within the maximum wait time.")))}),100)}));const t=await new Promise(((e,t)=>{if(MemberSpace.ready)try{e(window.MemberSpace.getMemberInfo())}catch(e){t("Error fetching MemberSpace member info: "+e)}else{const n=({detail:i})=>{try{e(i)}catch(e){t("Error in MemberSpace.ready event: "+e)}document.removeEventListener("MemberSpace.ready",n)};document.addEventListener("MemberSpace.ready",n)}}));t&&t.isLoggedIn&&(e.email=t.memberInfo.email,e.user_id=t.memberInfo.id)}catch(e){}return e}const o="https://api.colombochetty.com/",r=o+"api/authenticate",s=o+"auth/logout",c=o+"auth/refresh_token",l={tree:o+"api/tree/",individuals:o+"api/individuals/"},d=new class{async login(e,t){const n=await fetch(e,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(t),credentials:"include"});if(!n.ok)throw n}async logout(e){const t=await fetch(e,{method:"POST",credentials:"include"});if(!t.ok)throw t}async refreshToken(e){const t=this.getCookie("csrf_refresh_token"),n=await fetch(e,{method:"POST",credentials:"include",headers:{"X-CSRF-TOKEN-REFRESH":t}});if(!n.ok)throw n}async getRestData(e){const t=await fetch(e,{method:"GET",credentials:"include"});if(!t.ok)throw t;return await t.text()}async getGraphQLData(e,t){const n=this.getCookie("csrf_access_token"),i=await fetch(e,{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-TOKEN-ACCESS":n},body:JSON.stringify({query:t}),credentials:"include"});if(!i.ok)throw i;return await i.json()}getCookie(e){let t=null;if(document.cookie&&""!==document.cookie){const n=document.cookie.split(";");for(let i=0;i<n.length;i++){const a=n[i].trim();if(a.substring(0,e.length+1)===e+"="){t=decodeURIComponent(a.substring(e.length+1));break}}}return t}},h=new class{constructor(e,t,n,i,a,o){this.loginUrl=e,this.logoutUrl=n,this.refreshUrl=t,this.memberspace_tokenName=i,this.redirectPage=a,this.apiClient=o}async handleGraphQLError(e,t,n){try{if(this.isUnauthorizedError(e))return await this.handleUnauthorizedError(e),await this.apiClient.getGraphQLData(t,n);throw this.isSigFailure(e),await this.handleLogout(),e}catch(e){throw e}}async handleRestError(e,t){try{if(this.isUnauthorizedError(e))return await this.handleUnauthorizedError(e),await this.apiClient.getRestData(t);throw this.isSigFailure(e),await this.handleLogout(),e}catch(e){throw e}}async handleUnauthorizedError(e){const t=await e.json();if("Token has expired"===t.msg)await this.handleRefresh();else{if('Missing cookie "access_token_cookie"'!==t.msg)throw e;{const e=await a();await this.apiClient.login(this.loginUrl,e)}}}async handleRefresh(){try{await this.apiClient.refreshToken(this.refreshUrl)}catch(e){throw this.handleLogout(),e}}async handleLogout(){localStorage.removeItem(this.memberspace_tokenName),await this.apiClient.logout(this.logoutUrl),window.location.href=this.redirectPage}isUnauthorizedError(e){return e instanceof Response&&401===e.status}isSigFailure(e){return e instanceof Response&&422===e.status}}(r,c,s,"MemberSpaceWidget.token","/account",d);const u="tree",m="/person?id=";async function g(t,a=!0){try{const o=await async function(e,t){const n=l[e]+t;try{return await d.getRestData(n)}catch(e){return await h.handleRestError(e,n)}}(u,t);document.getElementById("tree-content").innerHTML=o,function(){i=!1,document.getElementsByClassName("root")[0];var t=document.getElementsByClassName("level-0")[0];t&&e(t.previousElementSibling,t,!1);var a,o=document.getElementsByClassName("collapsible");for(a=0;a<o.length;a++)o[a].addEventListener("click",(function(){n(this)}));var r=document.querySelector(".sticky-bar"),s=.5*window.innerHeight;window.addEventListener("scroll",(function(){window.pageYOffset>s?r.classList.remove("bar-hidden"):r.classList.add("bar-hidden")}));const c=function(e,t){let n;return function(){const i=this,a=arguments;clearTimeout(n),n=setTimeout((function(){n=null,e.apply(i,a)}),t)}}((function(){for(var t=document.querySelectorAll(".nested.active"),n=0;n<t.length;n++){var i=t[n];e(i.previousElementSibling,i,!1)}}),10);window.addEventListener("resize",c)}(),a?document.querySelectorAll(".tree-link").forEach((function(e){e.addEventListener("click",(function(){g(this.dataset.treeId)}))})):document.querySelectorAll(".tree-link").forEach((function(e){e.addEventListener("click",(function(){const e=m+this.dataset.treeId;window.location.href=e}))})),document.querySelectorAll(".profile-link").forEach((function(e){e.addEventListener("click",(function(){const e=m+this.dataset.profileId;window.location.href=e}))}))}catch(e){throw e}}window.initStaticTree=async function(e){try{await g(e,!1);document.getElementById("main-profile-link").style.display="none"}catch(e){}},window.activateAll=function(){document.getElementsByClassName("root")[0];for(var t=document.getElementsByClassName("level-0")[0],n=document.getElementsByClassName("nested"),a=0;a<n.length;a++){var o=n[a];i?o.classList.contains("level-0")||o.classList.remove("active"):o.classList.add("active")}if(i)e(t.previousElementSibling,t,!1);else for(a=0;a<n.length;a++){e((o=n[a]).previousElementSibling,o,!1)}var r=document.getElementsByClassName("collapsible");for(a=0;a<r.length;a++){var s=r[a];i?s.classList.remove("active"):s.classList.add("active")}i=!i}})();
+(()=>{"use strict";function e(t,n,i){const a=n.querySelectorAll(":scope > li"),o=a[a.length-1],r=o.getBoundingClientRect().height/2,s=o.getBoundingClientRect(),c=t.getBoundingClientRect(),l=s.bottom-c.bottom-r;n.style.setProperty("--ul-before-margin",`${l}px`);const d=n.parentElement;if(i&null!==d&"UL"===d.tagName){const t=d.previousElementSibling;null!==t&&e(t,d,!0)}}function t(e){for(var n=e.getElementsByClassName("nested"),i=0;i<n.length;i++)n[i].classList.remove("active"),t(n[i]);var a=e.getElementsByClassName("collapsible");for(i=0;i<a.length;i++)a[i].classList.remove("active")}function n(n){n.classList.toggle("active");const i=n.parentElement.parentElement.parentElement,a=i.nextElementSibling;a.classList.toggle("active"),e(i,a,!0),n.classList.contains("active")?n.textContent="":(n.textContent="",t(a))}var i=!1;async function a(){var e={email:"",user_id:""};try{await new Promise(((e,t)=>{let n=0;const i=setInterval((()=>{"undefined"!=typeof MemberSpace?(clearInterval(i),e()):(n+=100,n>=1e4&&(clearInterval(i),t("MemberSpace did not become available within the maximum wait time.")))}),100)}));const t=await new Promise(((e,t)=>{if(MemberSpace.ready)try{e(window.MemberSpace.getMemberInfo())}catch(e){t("Error fetching MemberSpace member info: "+e)}else{const n=({detail:i})=>{try{e(i)}catch(e){t("Error in MemberSpace.ready event: "+e)}document.removeEventListener("MemberSpace.ready",n)};document.addEventListener("MemberSpace.ready",n)}}));t&&t.isLoggedIn&&(e.email=t.memberInfo.email,e.user_id=t.memberInfo.id)}catch(e){}return e}const o="https://api.colombochetty.com/",r=o+"api/authenticate",s=o+"auth/logout",c=o+"auth/refresh_token",l={tree:o+"api/tree/",individuals:o+"api/individuals/"},d=new class{async login(e,t){const n=await fetch(e,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(t),credentials:"include"});if(!n.ok)throw n}async logout(e){const t=await fetch(e,{method:"POST",credentials:"include"});if(!t.ok)throw t}async refreshToken(e){const t=this.getCookie("csrf_refresh_token"),n=await fetch(e,{method:"POST",credentials:"include",headers:{"X-CSRF-TOKEN-REFRESH":t}});if(!n.ok)throw n}async getRestData(e){const t=await fetch(e,{method:"GET",credentials:"include"});if(!t.ok)throw t;return await t.text()}async getGraphQLData(e,t){const n=this.getCookie("csrf_access_token"),i=await fetch(e,{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-TOKEN-ACCESS":n},body:JSON.stringify({query:t}),credentials:"include"});if(!i.ok)throw i;return await i.json()}getCookie(e){let t=null;if(document.cookie&&""!==document.cookie){const n=document.cookie.split(";");for(let i=0;i<n.length;i++){const a=n[i].trim();if(a.substring(0,e.length+1)===e+"="){t=decodeURIComponent(a.substring(e.length+1));break}}}return t}},h=new class{constructor(e,t,n,i,a,o){this.loginUrl=e,this.logoutUrl=n,this.refreshUrl=t,this.memberspace_tokenName=i,this.redirectPage=a,this.apiClient=o}async handleGraphQLError(e,t,n){try{if(this.isUnauthorizedError(e))return await this.handleUnauthorizedError(e),await this.apiClient.getGraphQLData(t,n);throw this.isSigFailure(e),await this.handleLogout(),e}catch(e){throw e}}async handleRestError(e,t){try{if(this.isUnauthorizedError(e))return await this.handleUnauthorizedError(e),await this.apiClient.getRestData(t);throw this.isSigFailure(e),await this.handleLogout(),e}catch(e){throw e}}async handleUnauthorizedError(e){const t=await e.json();if("Token has expired"===t.msg)await this.handleRefresh();else{if('Missing cookie "access_token_cookie"'!==t.msg)throw e;{const e=await a();await this.apiClient.login(this.loginUrl,e)}}}async handleRefresh(){try{await this.apiClient.refreshToken(this.refreshUrl)}catch(e){throw this.handleLogout(),e}}async handleLogout(){localStorage.removeItem(this.memberspace_tokenName),await this.apiClient.logout(this.logoutUrl),window.location.href=this.redirectPage}isUnauthorizedError(e){return e instanceof Response&&401===e.status}isSigFailure(e){return e instanceof Response&&422===e.status}}(r,c,s,"MemberSpaceWidget.token","/account",d);const u="tree",m="/person?id=";async function g(t,a=!0){try{const o=await async function(e,t){const n=l[e]+t;try{return await d.getRestData(n)}catch(e){return await h.handleRestError(e,n)}}(u,t);document.getElementById("tree-content").innerHTML=o,function(){i=!1,document.getElementsByClassName("root")[0];var t=document.getElementsByClassName("level-0")[0];t&&e(t.previousElementSibling,t,!1);var a,o=document.getElementsByClassName("collapsible");for(a=0;a<o.length;a++)o[a].addEventListener("click",(function(){n(this)}));var r=document.querySelector(".sticky-bar"),s=.5*window.innerHeight;window.addEventListener("scroll",(function(){window.pageYOffset>s?r.classList.remove("bar-hidden"):r.classList.add("bar-hidden")}));const c=function(e,t){let n;return function(){const i=this,a=arguments;clearTimeout(n),n=setTimeout((function(){n=null,e.apply(i,a)}),t)}}((function(){for(var t=document.querySelectorAll(".nested.active"),n=0;n<t.length;n++){var i=t[n];e(i.previousElementSibling,i,!1)}}),10);window.addEventListener("resize",c)}(),a?document.querySelectorAll(".tree-link").forEach((function(e){e.addEventListener("click",(function(){g(this.dataset.treeId)}))})):document.querySelectorAll(".tree-link").forEach((function(e){e.addEventListener("click",(function(){const e=m+this.dataset.treeId;window.location.href=e}))})),document.querySelectorAll(".profile-link").forEach((function(e){e.addEventListener("click",(function(){const e=m+this.dataset.profileId;window.location.href=e}))}))}catch(e){throw e}}window.initStaticTree=async function(e){try{await g(e,!1);document.getElementById("main-profile-link").style.display="none"}catch(e){throw e}},window.activateAll=function(){document.getElementsByClassName("root")[0];for(var t=document.getElementsByClassName("level-0")[0],n=document.getElementsByClassName("nested"),a=0;a<n.length;a++){var o=n[a];i?o.classList.contains("level-0")||o.classList.remove("active"):o.classList.add("active")}if(i)e(t.previousElementSibling,t,!1);else for(a=0;a<n.length;a++){e((o=n[a]).previousElementSibling,o,!1)}var r=document.getElementsByClassName("collapsible");for(a=0;a<r.length;a++){var s=r[a];i?s.classList.remove("active"):s.classList.add("active")}i=!i}})();
