+++ docs/people.html
@@ -7,7 +7,6 @@ title: List of all people
 <meta name="robots" content="noindex, nofollow"><style>#__memberspace_modal_protected_page { position:fixed; top:0; left:0; width:100%; height:100%; background:#fff; z-index:2147483646; }</style><div id="__memberspace_modal_protected_page"></div>
 {% endif %}
 
-<script src="{{ site.baseurl }}/assets/js/table.js"></script>
 <script src="{{ site.baseurl }}/assets/js/all_people.js"></script>
 
 <link rel="stylesheet" href="{{ site.baseurl }}/assets/css/table_style.css" type="text/css">
