+++ webpack/js/person_routing.js
@@ -1,5 +1,30 @@
 import { setPersonData } from './requests/set_person';
 
+let PAGE_INIT = false;
+const SITE_BASE_URL = process.env.SITE_BASE_URL;
+
+document.addEventListener("DOMContentLoaded", async function() {
+    await initPerson();
+});
+
+document.addEventListener('MemberSpace.member.info', async function() {
+  	if (!PAGE_INIT) {
+	  	await initPerson();
+	}
+});
+
+async function initPerson() {
+	try {
+    	var path = window.location.pathname;
+  		const personId = getIdFromUrl();
+    	await setPersonData(personId);
+    	PAGE_INIT = true;  
+    } catch(error) {
+    	//
+    	console.log(error);
+    } 
+}
+
 function getIdFromUrl() {
     const url = new URL(window.location.href);
     const path = url.pathname;
@@ -14,15 +39,3 @@ function getIdFromUrl() {
     const pathParts = path.split('/');
     return pathParts[pathParts.length - 1];
 }
-
-const SITE_BASE_URL = process.env.SITE_BASE_URL;
-
-document.addEventListener("DOMContentLoaded", async function() {
-    var path = window.location.pathname;
-    console.log("hello")
-  	const personId = getIdFromUrl();
-    console.log("person id")
-    console.log(personId)
-    await setPersonData(personId);
-     
-});
