+++ webpack/js/utilities/tree_helpers.js
